plot.paircomp {psychotools}	R Documentation

Plotting Paired Comparison Data

Description

Plotting the frequency table from "paircomp" data.

Usage

  ## S3 method for class 'paircomp'
plot(x, off = 0.05,
  xlab = "Proportion of comparisons", ylab = "", tol.xlab = 0.05,
  abbreviate = TRUE, hue = NULL, chroma = 40, luminance = 80,
  xlim = c(0, 1), ylim = NULL, xaxs = "i", yaxs = "i", ...)

Arguments

x
an object of class "paircomp".

off
numeric. Offset between segments on the y-axis.

xlab, ylab
character. Axis labels.

tol.xlab
numeric. convenience tolerance parameter for x-axis annotation. If the distance between two labels drops under this threshold, they are plotted equidistantly.

abbreviate
logical or integer. Should object labels be abbreviated? Alternative an integer with the desired abbreviation length. The default is some heuristic based on the length of the labels.

hue
numeric. A vector of hues in [0, 360], recycled to the number of objects compared in x. A sequential palette is computed for each hue, see below.

chroma
numeric. Maximum chroma in the palette.

luminance
numeric. Minimum (and maximum) luminance in the palette. If omitted, the maximum is set to 95.

xlim, ylim, xaxs, yaxs, ...
graphical arguments passed to plot.

Details

The plot method creates a frequency table (using summary) and visualizes this using a sort of spine plot with HCL-based diverging palettes. See Zeileis, Hornik, Murrell (2009) for the underlying ideas.

References

Zeileis A, Hornik K, Murrell P (2009), Escaping RGBland: Selecting Colors for Statistical Graphics. Computational Statistics & Data Analysis, 53, 3259-3270. doi: 10.1016/j.csda.2008.11.033

See Also

paircomp

Examples

data("GermanParties2009", package = "psychotools")
par(mar = c(5, 6, 3, 6))
plot(GermanParties2009$preference, abbreviate = FALSE)

[Package psychotools version 0.7-6 Index]