subset.paircomp {psychotools}	R Documentation

Subsetting/Reordering Paired Comparison Data

Description

Selection of subsets of objects to be compared and/or reordering of objects in "paircomp" data.

Usage

  ## S3 method for class 'paircomp'
reorder(x, labels, ...)
  ## S3 method for class 'paircomp'
subset(x, subset, select, ...)

Arguments

x
an object of class "paircomp".

labels, select
character or integer. Either a vector of (at least two) elements of labels(x) or an integer with their position. Partial string matching is enabled.

subset
currently not implemented. (Should be a specification of subsets of subjects.)

...
currently not used.

Details

The subset method currently just calls the reorder method.

See Also

paircomp

Examples

pc <- paircomp(rbind(
  c(1,  1,  1), # a > b, a > c, b > c
  c(1,  1, -1), # a > b, a > c, b < c
  c(1, -1, -1), # a > b, a < c, b < c
  c(1,  1,  1)))
reorder(pc, c("c", "a"))

[Package psychotools version 0.7-6 Index]