worth {psychotools}	R Documentation

Extract Worth Parameters

Description

Generic functions for extracting worth parameters from paired comparison models.

Usage

  worth(object, ...)

Arguments

object
an object.

...
arguments passed to methods.

Details

Since version 0.3-0, calls to worth are internally passed over to itempar.

See Also

btmodel, raschmodel

Examples

o <- options(digits = 4)

## data
data("GermanParties2009", package = "psychotools")

## Bradley-Terry model
bt <- btmodel(GermanParties2009$preference)

## worth parameters
worth(bt)

## or
itempar(bt)

options(digits = o$digits)

[Package psychotools version 0.7-6 Index]